replay buffer start/stop button
Add a button to start/stop the replay buffer via the remote. Still doesn't sync up with OBS's actual replay buffer status so the status isn't to be relied upon for the replay buffer status but the save button will always try to save regardless of whether the buffer is active or not.